# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API pour la gestion des utilisateurs et l'authentification.",
        "contact": {
            "email": "sahliabderazak530@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addOffres": {
            "post": {
                "tags": [
                    "Offres"
                ],
                "summary": "Ajouter une offre",
                "operationId": "26456097db36cdba2b2845aa17202922",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "departement",
                                    "poste",
                                    "description",
                                    "dateExpiration"
                                ],
                                "properties": {
                                    "departement": {
                                        "type": "string",
                                        "example": "Informatique"
                                    },
                                    "poste": {
                                        "type": "string",
                                        "example": "Développeur Backend"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Développement en Laravel"
                                    },
                                    "dateExpiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre ajoutée avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
        "/api/offres-departement": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Récupérer les offres selon le département de l'utilisateur connecté",
                "operationId": "6d43d2b77df7a932d57227363688bad3",
                "responses": {
                    "200": {
                        "description": "Liste des offres du département"
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/validerOffre/{id}": {
            "put": {
                "tags": [
                    "Offres"
                ],
                "summary": "Valider une offre par son ID",
                "operationId": "f8cba03a17557ccfc6a395d6ce43c3d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à valider",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre validée avec succès"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/supprimerOffre/{id}": {
            "delete": {
                "tags": [
                    "Offres"
                ],
                "summary": "Supprimer une offre par son ID",
                "operationId": "cfa40483e38c84b3268a2a4c7eec3d25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée avec succès"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/offres-departement/{id}": {
            "put": {
                "tags": [
                    "Offres"
                ],
                "summary": "Modifier une offre non validée",
                "operationId": "9fd8e5b682664ca4c6ae778910561dd8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departement": {
                                        "type": "string",
                                        "example": "RH"
                                    },
                                    "poste": {
                                        "type": "string",
                                        "example": "Manager"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Gestion des équipes"
                                    },
                                    "dateExpiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre modifiée avec succès"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/prolongerOffre/{id}": {
            "put": {
                "tags": [
                    "Offres"
                ],
                "summary": "Prolonger la date d'expiration d'une offre validée",
                "operationId": "aa95cbd07ce77ec6e1f69846f7359b8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à prolonger",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dateExpiration"
                                ],
                                "properties": {
                                    "dateExpiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "La date d'expiration de l'offre a été prolongée avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation ou l'offre n'est pas validée"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/offres-expirees": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Afficher les offres expirées pour le département du recruteur",
                "operationId": "8aa81bb81a6a03564e24ff85258ec56b",
                "responses": {
                    "200": {
                        "description": "Liste des offres expirées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Offre"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Le département de l'utilisateur n'est pas défini"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Offre": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "departement": {
                        "type": "string"
                    },
                    "poste": {
                        "type": "string"
                    },
                    "datePublication": {
                        "type": "string",
                        "format": "date"
                    },
                    "dateExpiration": {
                        "type": "string",
                        "format": "date"
                    },
                    "valider": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    },
...
}